fix running locally with only couchdb env vars
