+++ package.json
@@ -26,9 +26,9 @@
     "body-parser": "^1.20.1",
     "cookie-session": "^2.0.0",
     "crypto-js": "^4.1.1",
+    "dhp-auth-lib": "github:digitalhealthpass/dhp-auth-lib",
     "ecdsa-secp256r1": "^1.3.3",
     "express": "^4.18.1",
-    "healthpass-auth-lib": "github:WH-HealthPass/healthpass-auth-lib",
     "helmet": "^6.0.0",
     "https": "^1.0.0",
     "ibmcloud-appid": "^6.2.6",
